refactor: use unified toast notifier for trash and conversion success messages
