+++ packages/app-desktop/gui/ConversionNotification/ConversionNotification.tsx
@@ -2,8 +2,7 @@ import * as React from 'react';
 import { useEffect } from 'react';
 import { _ } from '@joplin/lib/locale';
 import { Dispatch } from 'redux';
-import { NotificationType } from '../PopupNotification/types';
-import useToastNotifier from '../../gui/hooks/useToast';
+import useToastNotifier from '../hooks/useToastNotifier';
 
 interface Props {
 	noteId: string;
@@ -18,13 +17,7 @@ export default (props: Props) => {
 
 		props.dispatch({ type: 'NOTE_HTML_TO_MARKDOWN_DONE', value: '' });
 
-		notify(
-			_('The note has been converted to Markdown and the original note has been moved to the trash'),
-			{
-				type: NotificationType.Success,
-				delay: 4000,
-			},
-		);
+		notify(_('The note has been converted to Markdown and the original note has been moved to the trash.'));
 	}, [props.dispatch, props.noteId, notify]);
 
 	return <div style={{ display: 'none' }}/>;
